refactor: extract hook for rescale event
