+++ typescript/packages/subsurface-viewer/src/components/Map.tsx
@@ -72,6 +72,7 @@ import IntersectionView from "../views/intersectionView";
 
 import type { LightsType, TLayerDefinition } from "../SubsurfaceViewer";
 import { getZoom, useLateralZoom } from "../utils/camera";
+import { useHandleRescale } from "../utils/event";
 
 export type { BoundingBox2D, BoundingBox3D };
 /**
@@ -88,58 +89,6 @@ const minZoom3D = -12;
 const maxZoom3D = 12;
 const minZoom2D = -12;
 const maxZoom2D = 4;
-
-// https://developer.mozilla.org/docs/Web/API/KeyboardEvent
-type ArrowEvent = {
-    key: "ArrowUp" | "ArrowDown" | "PageUp" | "PageDown";
-    shiftModifier: boolean;
-};
-
-function updateZScaleReducer(zScale: number, action: ArrowEvent): number {
-    return zScale * getZScaleModifier(action);
-}
-
-function getZScaleModifier(arrowEvent: ArrowEvent): number {
-    let scaleFactor = 0;
-    switch (arrowEvent.key) {
-        case "ArrowUp":
-            scaleFactor = 0.05;
-            break;
-        case "ArrowDown":
-            scaleFactor = -0.05;
-            break;
-        case "PageUp":
-            scaleFactor = 0.25;
-            break;
-        case "PageDown":
-            scaleFactor = -0.25;
-            break;
-        default:
-            break;
-    }
-    if (arrowEvent.shiftModifier) {
-        scaleFactor /= 5;
-    }
-    return 1 + scaleFactor;
-}
-
-function convertToArrowEvent(event: KeyboardEvent): ArrowEvent | null {
-    if (event.type === "keydown") {
-        switch (event.key) {
-            case "ArrowUp":
-            case "ArrowDown":
-            case "PageUp":
-            case "PageDown":
-                return {
-                    key: event.key,
-                    shiftModifier: event.shiftKey,
-                };
-            default:
-                return null;
-        }
-    }
-    return null;
-}
 
 function parseLights(lights?: LightsType): LightingEffect[] | undefined {
     if (!lights) {
@@ -258,7 +207,7 @@ export interface ViewportType {
     zoom?: number;
     rotationX?: number;
     rotationOrbit?: number;
-    verticalScale?: number;
+    verticalScale?: number | (() => number);
 
     isSync?: boolean;
 }
@@ -505,7 +454,7 @@ const Map: React.FC<MapProps> = ({
     );
 
     // Used for scaling in z direction using arrow keys.
-    const [zScale, updateZScale] = React.useReducer(updateZScaleReducer, 1);
+    const [zScale, divRef, shiftHeld] = useHandleRescale();
 
     // compute the viewport margins
     const viewPortMargins = React.useMemo<MarginsType>(() => {
@@ -547,39 +496,6 @@ const Map: React.FC<MapProps> = ({
     // multiple well layers
     const [multipleWells, setMultipleWells] = useState<string[]>([]);
     const [selectedWell, setSelectedWell] = useState<string>("");
-    const [shiftHeld, setShiftHeld] = useState(false);
-
-    const divRef = React.useRef(null);
-
-    useEffect(() => {
-        const keyDownHandler = (e: KeyboardEvent) => {
-            const arrowEvent = convertToArrowEvent(e);
-            if (arrowEvent) {
-                updateZScale(arrowEvent);
-                // prevent being handled by regular OrbitController
-                e.stopPropagation();
-            }
-            if (e.key === "Shift") {
-                setShiftHeld(true);
-            }
-        };
-        const keyUpHandler = (e: KeyboardEvent) => {
-            if (e.key === "Shift") {
-                setShiftHeld(false);
-            }
-        };
-
-        // eslint-disable-next-line @typescript-eslint/no-explicit-any
-        const element = divRef.current as any;
-
-        element?.addEventListener("keydown", keyDownHandler, true);
-        element?.addEventListener("keyup", keyUpHandler, true);
-
-        return () => {
-            element?.removeEventListener("keydown", keyDownHandler);
-            element?.removeEventListener("keyup", keyUpHandler);
-        };
-    }, [updateZScale, setShiftHeld]);
 
     useEffect(() => {
         const layers = deckRef.current?.deck?.props.layers;
