+++ typescript/packages/subsurface-viewer/src/storybook/examples/CameraControlExamples.stories.tsx
@@ -574,6 +574,48 @@ export const ScaleY: StoryObj<typeof ScaleYComponent> = {
     render: (args) => <ScaleYComponent {...args} />,
 };
 
+const ScaleVertical3dComponent = ({
+    verticalScale,
+}: {
+    verticalScale: number;
+}) => {
+    const viewerProps: SubsurfaceViewerProps = {
+        id: "ScaleY",
+        bounds: [-10, -10, 10, 10],
+        layers: [AXES2D, SQUARE_GEOMETRY_LAYER],
+        views: {
+            layout: [1, 1],
+            viewports: [
+                {
+                    id: "section",
+                    verticalScale,
+                    zoom: 2,
+                    show3D: true,
+                },
+            ],
+        },
+    };
+    return <SubsurfaceViewer {...viewerProps} />;
+};
+
+export const ScaleVertical3d: StoryObj<typeof ScaleVertical3dComponent> = {
+    args: { verticalScale: 1.5 },
+    argTypes: {
+        verticalScale: {
+            control: { type: "range", min: -1, max: 10, step: 0.1 },
+        },
+    },
+    parameters: {
+        docs: {
+            ...defaultStoryParameters.docs,
+            description: {
+                story: "Vertical scaling example in panoramic view.",
+            },
+        },
+    },
+    render: (args) => <ScaleVertical3dComponent {...args} />,
+};
+
 const ScaleYWithCameraPositionComponent = ({
     verticalScale,
 }: {
